new direction
I decided to go in a new direction, with this project and this branch is
the start of it.  First I am attempting to use few to no graphics, so I
am challenging myself to use CSS to make my graphics and keep them
simple.  Second, I went with a different approach to editing the items.
The current state of this is rough at best, but will continue to improve
over the next week.